pub use glyph::Error as GlyphError
since `mod glyph;` is private, re-export its `pub enum Error`
alongside the re-exported `GlyphCache`. Otherwise, it is impossible
for someone to write wrapper functions that return `Result<GlyphCache,
glyph::Error>`, because the error's parent module is private.

(You may want to also rename the definition of `glyph::Error` itself,
because it may improve the doc presentation of the situation.)